% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_collection_list.R
\name{new_xlsx_collection_item}
\alias{new_xlsx_collection_item}
\title{Create an xlsx collection item}
\usage{
new_xlsx_collection_item(st, caption, sheet_name, footer = NULL)
}
\arguments{
\item{st}{A \link[styledTables:StyledTable-styled_table-method]{StyledTable} class ojbect, which
should be stored.}

\item{caption}{A caption string for the table.}

\item{sheet_name}{A string used as sheet name, when the table is written
to an xlsx file.}

\item{footer}{A string used as table footer.}
}
\description{
This create a \code{xlsx_collection_item} class object, which can be used
in order to store a singel StyledTable together with caption,
footer and sheet name.
}
