% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_collection_list.R
\name{new_xlsx_collection_list}
\alias{new_xlsx_collection_list}
\title{Create an \code{xlsx_collection_list} class object.}
\usage{
new_xlsx_collection_list(obj)
}
\arguments{
\item{obj}{A list holding the items.}
}
\value{
A \code{xlsx_collection_list} class object.
}
\description{
These class objects bundle mutliple
\link[=new_xlsx_collection_item]{xlsx_collection_item} in a single
kind of list class object.
}
