% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_collection_list.R
\name{validate_xlsx_collection_list}
\alias{validate_xlsx_collection_list}
\title{Validate the \link[=new_xlsx_collection_list]{xlsx_collection_list} class object}
\usage{
validate_xlsx_collection_list(
  obj,
  validate_class = TRUE,
  err_h = composerr("Invalid 'xlsx_collection_list'")
)
}
\arguments{
\item{obj}{The object, which should be validated.}

\item{validate_class}{A logical flag, defining if the class string should
also be evaluated.}

\item{err_h}{An error handling function.}
}
\description{
Validate the \link[=new_xlsx_collection_list]{xlsx_collection_list} class object
}
