% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxcollection_item.R, R/xlsxcollection_list.R
\name{validate_xlsxcollection_item}
\alias{validate_xlsxcollection_item}
\title{Validate the \link[=new_xlsxcollection_item]{xlsxcollection_item} class object}
\usage{
validate_xlsxcollection_item(
  obj,
  validate_class = TRUE,
  err_h = composerr("Invalid 'xlsxcollection_item'")
)

validate_xlsxcollection_item(
  obj,
  validate_class = TRUE,
  err_h = composerr("Invalid 'xlsxcollection_item'")
)
}
\arguments{
\item{obj}{The object, which should be validated.}

\item{validate_class}{A logical flag, defining if the class string should
also be evaluated.}

\item{err_h}{An error handler created with \link[composerr:composerr]{composerr}}
}
\description{
Validate the \link[=new_xlsxcollection_item]{xlsxcollection_item} class object

Validate the \link[=new_xlsxcollection_item]{xlsxcollection_item} class object
}
