% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxcollection_list.R
\name{validate_xlsxcollection_list}
\alias{validate_xlsxcollection_list}
\title{Validate the \link[=new_xlsxcollection_list]{xlsxcollection_list} class object}
\usage{
validate_xlsxcollection_list(
  obj,
  validate_class = TRUE,
  err_h = composerr("Invalid 'xlsxcollection_list'")
)
}
\arguments{
\item{obj}{The object, which should be validated.}

\item{validate_class}{A logical flag, defining if the class string should
also be evaluated.}

\item{err_h}{An error handler created with \link[composerr:composerr]{composerr}}
}
\description{
Validate the \link[=new_xlsxcollection_list]{xlsxcollection_list} class object
}
