% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{xlsx_collection_append_table}
\alias{xlsx_collection_append_table}
\title{Append a \link[=new_xlsx_collection_item]{xlsx_collection_item}
class object to an \link[xlsx:Workbook]{xlsx-workbook}}
\usage{
xlsx_collection_append_table(wb, sheet_name, st, caption, footer = NULL)
}
\arguments{
\item{wb}{An \link[xlsx:Workbook]{xlsx-workbook}}

\item{sheet_name}{A string used as sheet name for the table}

\item{st}{A \link[styledTables:StyledTable-styled_table-method]{StyledTable} class object.}

\item{caption}{A string used as table caption.}

\item{footer}{An optional string used as table footer.}
}
\description{
Append a \link[=new_xlsx_collection_item]{xlsx_collection_item}
class object to an \link[xlsx:Workbook]{xlsx-workbook}
}
\seealso{
\code{\link[=xlsx_collection_list_stored_tables]{xlsx_collection_list_stored_tables()}},
\code{\link[=xlsx_collection_init_store_table]{xlsx_collection_init_store_table()}},
\code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}},
\code{\link[=xlsx_collection_use_latex_table_counter]{xlsx_collection_use_latex_table_counter()}},
\code{\link[=xlsx_collection_create_excel]{xlsx_collection_create_excel()}},
\code{\link[=xlsx_collection_append_toc]{xlsx_collection_append_toc()}}
}
