% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{xlsx_collection_append_toc}
\alias{xlsx_collection_append_toc}
\title{Append a table of contents sheet to the xlsx collection file}
\usage{
xlsx_collection_append_toc(wb, xlsx_collection_list, toc_header)
}
\arguments{
\item{wb}{An \link[xlsx:Workbook]{xlsx-workbook}}

\item{xlsx_collection_list}{An
\link[=new_xlsx_collection_list]{xlsx_collection_list} class object holding
the stored tables.}

\item{toc_header}{A string used as caption for the table of contents.}
}
\description{
Append a table of contents sheet to the xlsx collection file
}
\seealso{
\code{\link[=xlsx_collection_list_stored_tables]{xlsx_collection_list_stored_tables()}},
\code{\link[=xlsx_collection_init_store_table]{xlsx_collection_init_store_table()}},
\code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}},
\code{\link[=xlsx_collection_use_latex_table_counter]{xlsx_collection_use_latex_table_counter()}},
\code{\link[=xlsx_collection_create_excel]{xlsx_collection_create_excel()}},
\code{\link[=xlsx_collection_append_table]{xlsx_collection_append_table()}}
}
