% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{xlsx_collection_create_excel}
\alias{xlsx_collection_create_excel}
\title{Create an xlsx collection}
\usage{
xlsx_collection_create_excel(xlsx_collection_list, xlsx_path, toc_header)
}
\arguments{
\item{xlsx_collection_list}{A \link[=new_xlsx_collection_list]{xlsx_collection_list}
class object, usually created by \code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}}.}

\item{xlsx_path}{File path of the xlsx file.}

\item{toc_header}{A string used as caption for the table of contents.}
}
\value{
The file path given in \code{xlsx_path}.
}
\description{
This function writes the read tables to an xlsx file.
}
\seealso{
\code{\link[=xlsx_collection_list_stored_tables]{xlsx_collection_list_stored_tables()}},
\code{\link[=xlsx_collection_init_store_table]{xlsx_collection_init_store_table()}},
\code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}},
\code{\link[=xlsx_collection_use_latex_table_counter]{xlsx_collection_use_latex_table_counter()}},
\code{\link[=xlsx_collection_append_table]{xlsx_collection_append_table()}},
\code{\link[=xlsx_collection_append_toc]{xlsx_collection_append_toc()}}
}
