% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tables.R
\name{xlsx_collection_init_store_table}
\alias{xlsx_collection_init_store_table}
\title{Initialize the function used for storing the tables.}
\usage{
xlsx_collection_init_store_table(storage_dir)
}
\arguments{
\item{storage_dir}{A string holding the path to a temporary folder where
the \link[styledTables:StyledTable-styled_table-method]{StyledTable} objects should be stored.}
}
\value{
A function, used for storing \link[styledTables:StyledTable-styled_table-method]{styledTable}
objects. With each call of the function a single table is stored. The function
takes the following arguments:
\itemize{
\item \code{st}: The \link[styledTables:StyledTable-styled_table-method]{StyledTable} class object,
which should be stored.
\item \code{sheet_name}: A string, which should be used as xlsx sheet name.
In case this tables origin from a LaTeX table and you want to use
the LaTeX counter later on, you must pass the LaTeX label of the table
to the \code{sheet_name} argument. When you call
\code{\link[=xlsx_collection_use_latex_table_counter]{xlsx_collection_use_latex_table_counter()}} the stored \code{sheet_name} property
will be compared with the LaTeX labels found in the \code{aux}-file of the
LaTeX report.
\item \code{caption}: A string, which should be used as table caption.
\item \code{footer}: A string, which should be used as table footer.
}
}
\description{
Initialize the function used for storing the tables.
}
\seealso{
\code{\link[=xlsx_collection_list_stored_tables]{xlsx_collection_list_stored_tables()}},
\code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}},
\code{\link[=xlsx_collection_use_latex_table_counter]{xlsx_collection_use_latex_table_counter()}},
\code{\link[=xlsx_collection_create_excel]{xlsx_collection_create_excel()}}
}
