% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tables.R
\name{xlsx_collection_list_stored_tables}
\alias{xlsx_collection_list_stored_tables}
\title{List all stored tables}
\usage{
xlsx_collection_list_stored_tables(storage_dir)
}
\arguments{
\item{storage_dir}{A string holding the path to a temporary folder where
the \link[styledTables:StyledTable-styled_table-method]{StyledTable} objects should be stored.}
}
\value{
A character vector holding the file paths of all tables stored in
the directory.
}
\description{
List all stored tables
}
\seealso{
\code{\link[=xlsx_collection_init_store_table]{xlsx_collection_init_store_table()}},
\code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}},
\code{\link[=xlsx_collection_use_latex_table_counter]{xlsx_collection_use_latex_table_counter()}},
\code{\link[=xlsx_collection_create_excel]{xlsx_collection_create_excel()}}
}
