% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tables.R
\name{xlsx_collection_read_stored_tables}
\alias{xlsx_collection_read_stored_tables}
\title{Read multiple stored tables}
\usage{
xlsx_collection_read_stored_tables(stored_tables)
}
\arguments{
\item{stored_tables}{A character vector holding the file paths to all
stored tables, which should be read.}
}
\value{
A \link[=new_xlsx_collection_list]{xlsx_collection_list} class object,
holding all stored tables.
}
\description{
Read multiple stored tables
}
\seealso{
\code{\link[=xlsx_collection_list_stored_tables]{xlsx_collection_list_stored_tables()}},
\code{\link[=xlsx_collection_init_store_table]{xlsx_collection_init_store_table()}},
\code{\link[=xlsx_collection_use_latex_table_counter]{xlsx_collection_use_latex_table_counter()}},
\code{\link[=xlsx_collection_create_excel]{xlsx_collection_create_excel()}}
}
