% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tables.R
\name{xlsx_collection_use_latex_table_counter}
\alias{xlsx_collection_use_latex_table_counter}
\title{Update the xlsx-collection by using the generated LaTeX table counter.}
\usage{
xlsx_collection_use_latex_table_counter(xlsx_collection_list, aux_path)
}
\arguments{
\item{xlsx_collection_list}{A \link[=new_xlsx_collection_list]{xlsx_collection_list}
class object, usually created by \code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}}.}

\item{aux_path}{A string holding the path to the \verb{*.aux} file generated
by LaTeX when calling \code{pdflatex} (or \code{lualatex} etc.).}
}
\value{
A modified  \link[=new_xlsx_collection_list]{xlsx_collection_list}
class object.
}
\description{
This function is only needed when the stored tables were part of a generated
LaTeX report (rnw-file) and the numbering of resulting pdf (created by LaTeX)
should also be used for the xlsx collection.
This function modifies an
\link[=new_xlsx_collection_list]{xlsx_collection_list} class object, such that
the defined sheet names and sheet ordering uses the table counter
created by LaTeX.
}
\seealso{
\code{\link[=xlsx_collection_list_stored_tables]{xlsx_collection_list_stored_tables()}},
\code{\link[=xlsx_collection_init_store_table]{xlsx_collection_init_store_table()}},
\code{\link[=xlsx_collection_read_stored_tables]{xlsx_collection_read_stored_tables()}},
\code{\link[=xlsx_collection_create_excel]{xlsx_collection_create_excel()}}
}
