% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{xlsxcollection_append_toc}
\alias{xlsxcollection_append_toc}
\title{Append a table of contents sheet to the xlsx collection file}
\usage{
xlsxcollection_append_toc(
  wb,
  xlsxcollection_list,
  toc_caption = NULL,
  toc_format = NULL
)
}
\arguments{
\item{wb}{An \link[xlsx:Workbook]{xlsx-workbook}}

\item{xlsxcollection_list}{An
\link[=new_xlsxcollection_list]{xlsxcollection_list} class object holding
the stored tables.}

\item{toc_caption}{An optional string used as caption for the table of contents.}

\item{toc_format}{An optional function applying \code{styledTables} formatting
commands to the \link[styledTables:StyledTable-styled_table-method]{StyledTable} class object
holding the entire toc-table. If omitted, then no additional formatting
is applied.}
}
\description{
Append a table of contents sheet to the xlsx collection file
}
\seealso{
\code{\link[=xlsxcollection_list_stored_tables]{xlsxcollection_list_stored_tables()}},
\code{\link[=xlsxcollection_init_store_table]{xlsxcollection_init_store_table()}},
\code{\link[=xlsxcollection_read_stored_tables]{xlsxcollection_read_stored_tables()}},
\code{\link[=xlsxcollection_use_latex_table_counter]{xlsxcollection_use_latex_table_counter()}},
\code{\link[=xlsxcollection_create_excel]{xlsxcollection_create_excel()}},
\code{\link[=xlsxcollection_append_table]{xlsxcollection_append_table()}}
}
