% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_excel.R
\name{xlsxcollection_create_excel}
\alias{xlsxcollection_create_excel}
\title{Create an xlsx collection}
\usage{
xlsxcollection_create_excel(
  xlsxcollection_list,
  xlsx_path,
  toc_caption = NULL,
  toc_format = NULL,
  toc_include = TRUE
)
}
\arguments{
\item{xlsxcollection_list}{A \link[=new_xlsxcollection_list]{xlsxcollection_list}
class object, usually created by \code{\link[=xlsxcollection_read_stored_tables]{xlsxcollection_read_stored_tables()}}.}

\item{xlsx_path}{File path of the xlsx file.}

\item{toc_caption}{An optional string used as caption for the table of contents.}

\item{toc_format}{An optional function applying \code{styledTables} formatting
commands to the \link[styledTables:StyledTable-styled_table-method]{StyledTable} class object
holding the entire toc-table. If omitted, then no additional formatting
is applied.}

\item{toc_include}{An logical flag, defining if the resulting excel file
should have a table of contents as first sheet.}
}
\value{
The file path given in \code{xlsx_path}.
}
\description{
This function writes the read tables to an xlsx file.
}
\seealso{
\code{\link[=xlsxcollection_list_stored_tables]{xlsxcollection_list_stored_tables()}},
\code{\link[=xlsxcollection_init_store_table]{xlsxcollection_init_store_table()}},
\code{\link[=xlsxcollection_read_stored_tables]{xlsxcollection_read_stored_tables()}},
\code{\link[=xlsxcollection_use_latex_table_counter]{xlsxcollection_use_latex_table_counter()}},
\code{\link[=xlsxcollection_append_table]{xlsxcollection_append_table()}},
\code{\link[=xlsxcollection_append_toc]{xlsxcollection_append_toc()}}
}
