% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tables.R
\name{xlsxcollection_list_stored_tables}
\alias{xlsxcollection_list_stored_tables}
\title{List all stored tables}
\usage{
xlsxcollection_list_stored_tables(storage_dir)
}
\arguments{
\item{storage_dir}{A string holding the path to a temporary folder where
the \link[styledTables:StyledTable-styled_table-method]{StyledTable} objects should be stored.}
}
\value{
A character vector holding the file paths of all tables stored in
the directory.
}
\description{
List all stored tables
}
\seealso{
\code{\link[=xlsxcollection_init_store_table]{xlsxcollection_init_store_table()}},
\code{\link[=xlsxcollection_read_stored_tables]{xlsxcollection_read_stored_tables()}},
\code{\link[=xlsxcollection_use_latex_table_counter]{xlsxcollection_use_latex_table_counter()}},
\code{\link[=xlsxcollection_create_excel]{xlsxcollection_create_excel()}}
}
