% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tables.R
\name{xlsxcollection_read_stored_tables}
\alias{xlsxcollection_read_stored_tables}
\title{Read multiple stored tables}
\usage{
xlsxcollection_read_stored_tables(stored_tables)
}
\arguments{
\item{stored_tables}{A character vector holding the file paths to all
stored tables, which should be read.}
}
\value{
A \link[=new_xlsxcollection_list]{xlsxcollection_list} class object,
holding all stored tables.
}
\description{
Read multiple stored tables
}
\seealso{
\code{\link[=xlsxcollection_list_stored_tables]{xlsxcollection_list_stored_tables()}},
\code{\link[=xlsxcollection_init_store_table]{xlsxcollection_init_store_table()}},
\code{\link[=xlsxcollection_use_latex_table_counter]{xlsxcollection_use_latex_table_counter()}},
\code{\link[=xlsxcollection_create_excel]{xlsxcollection_create_excel()}}
}
