% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tables.R
\name{xlsxcollection_use_latex_table_counter}
\alias{xlsxcollection_use_latex_table_counter}
\title{Update the xlsx-collection by using the generated LaTeX table counter.}
\usage{
xlsxcollection_use_latex_table_counter(xlsxcollection_list, aux_path)
}
\arguments{
\item{xlsxcollection_list}{A \link[=new_xlsxcollection_list]{xlsxcollection_list}
class object, usually created by \code{\link[=xlsxcollection_read_stored_tables]{xlsxcollection_read_stored_tables()}}.}

\item{aux_path}{A string holding the path to the \verb{*.aux} file generated
by LaTeX when calling \code{pdflatex} (or \code{lualatex} etc.).}
}
\value{
A modified  \link[=new_xlsxcollection_list]{xlsxcollection_list}
class object.
}
\description{
This function is only needed when the stored tables were part of a generated
LaTeX report (rnw-file) and the numbering of resulting pdf (created by LaTeX)
should also be used for the xlsx collection.
This function modifies an
\link[=new_xlsxcollection_list]{xlsxcollection_list} class object, such that
the defined sheet names and sheet ordering uses the table counter
created by LaTeX.
}
\seealso{
\code{\link[=xlsxcollection_list_stored_tables]{xlsxcollection_list_stored_tables()}},
\code{\link[=xlsxcollection_init_store_table]{xlsxcollection_init_store_table()}},
\code{\link[=xlsxcollection_read_stored_tables]{xlsxcollection_read_stored_tables()}},
\code{\link[=xlsxcollection_create_excel]{xlsxcollection_create_excel()}}
}
