% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_collection_list.R
\name{new_xlsx_collection_list}
\alias{new_xlsx_collection_list}
\title{Create an \code{xlsx_collection_list} class object.}
\usage{
new_xlsx_collection_list(obj, ...)
}
\arguments{
\item{obj}{One can eather pass a list of items or a singe item.}

\item{...}{additional \link[=new_xlsx_collection_item]{xlsx_collection_item} class objects.}
}
\value{
A \code{xlsx_collection_list} class object holding all items.
}
\description{
Bundles mutliple
\link[=new_xlsx_collection_item]{xlsx_collection_item} into a single
object of class \code{xlsx_collection_list}.
}
